Add files via upload
Divided the code into files, added processing of incorrect login or password input
